requirements.txt
discord.py>=2.3.2
aiofiles>=23.2.1
aiosqlite>=0.19.0
Pillow>=10.0.1
python-dotenv>=1.0.0
schedule>=1.2.0
asyncio-cron>=0.0.2
colorama>=0.4.6
typing-extensions>=4.8.0